migrate to version 2
Okay... I'll be honest, I must have been completely insane when I first wrote
this. Coming back to it, there are so many aparrent design flaws and mistakes
the only solution I could think of would be to essentially rewrite it from
scratch ~~and that's sort of what I've done~~.

I'd like to think I've grown as a programmer, and that once I hopefully do so
again, I'll look back and be much prouder of this version.

An overview of important changes:

* __init__.py has been properly utilised
now you can do
  `from download_queue import DownloadQueue`
instead of having to fully quantify it as:
  `from download_queue.download_queue import DownloadQueue`.

* deprecated the overflow queue
now if you try to add an item and the queue is already full, your calling
thread will be blocked until your request can be satisfied. The overflow
queue did make sense to me at the time, but in retrospect I just ended up
manually waiting for the queue to be free so I could log that I've moved
on to the next item.

* switched to synchronising primitives instead of polling
I had initially intended to do this from the start, in fact polling was just
a rushed solution while I focused on other things, but in the end I never got
round to it. Now I have. This should give the queue a substantial performance
boost.

* merged all the decorators used in downloader into one decorator
I did initially want to keep them seperated because they offered differing
functionality, but that just ended up making them bloated.

* implemented propper logging directly from DownloadQueue
I wasn't all that used to logging back when I first wrote this, but in
retrospect its essential to this.

* added some unittests for GenericDownload
once I can think of how to test download queue I'll write some for it as well.

Signed-off-by: MoHKale <mohkalsin@gmail.com>